feat(auth): minor fixes for groups and add tests

+++ libs/core/src/main/kotlin/io/stereov/singularity/auth/group/service/GroupService.kt
@@ -3,8 +3,9 @@ package io.stereov.singularity.auth.group.service
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.stereov.singularity.auth.core.service.AuthorizationService
 import io.stereov.singularity.auth.group.dto.request.CreateGroupRequest
-import io.stereov.singularity.auth.group.dto.response.UpdateGroupRequest
+import io.stereov.singularity.auth.group.dto.request.UpdateGroupRequest
 import io.stereov.singularity.auth.group.exception.model.GroupKeyExistsException
+import io.stereov.singularity.auth.group.exception.model.InvalidGroupTranslationException
 import io.stereov.singularity.auth.group.mapper.GroupMapper
 import io.stereov.singularity.auth.group.model.GroupDocument
 import io.stereov.singularity.auth.group.model.GroupTranslation
@@ -13,19 +14,22 @@ import io.stereov.singularity.database.core.service.TranslatableCrudService
 import io.stereov.singularity.global.exception.model.DocumentNotFoundException
 import io.stereov.singularity.global.properties.AppProperties
 import io.stereov.singularity.user.core.model.Role
+import io.stereov.singularity.user.core.service.UserService
 import jakarta.annotation.PostConstruct
+import kotlinx.coroutines.reactive.collect
+import kotlinx.coroutines.reactor.asFlux
 import kotlinx.coroutines.runBlocking
-import org.bson.types.ObjectId
 import org.springframework.data.mongodb.core.ReactiveMongoTemplate
 import org.springframework.stereotype.Service
 
 @Service
 class GroupService(
-    private val repository: GroupRepository,
+    override val repository: GroupRepository,
     override val appProperties: AppProperties,
     private val authorizationService: AuthorizationService,
     override val reactiveMongoTemplate: ReactiveMongoTemplate,
-    private val groupMapper: GroupMapper
+    private val groupMapper: GroupMapper,
+    private val userService: UserService
 ) : TranslatableCrudService<GroupTranslation, GroupDocument> {
 
     override val logger = KotlinLogging.logger {}
@@ -55,22 +59,15 @@ class GroupService(
     private suspend fun createNotAuthorized(req: CreateGroupRequest): GroupDocument {
         logger.debug { "Creating group with key \"${req.key}\"" }
 
+        if (req.translations.isEmpty())
+            throw InvalidGroupTranslationException("Failed to create group: at least one translation is needed")
+
+        if (!req.translations.containsKey(appProperties.locale))
+            throw InvalidGroupTranslationException("Failed to create group: default locale ${appProperties.locale} is not contained in translations")
 
         if (existsByKey(req.key)) throw GroupKeyExistsException(req.key)
 
         return save(groupMapper.createGroup(req))
-    }
-
-    suspend fun save(group: GroupDocument): GroupDocument {
-        logger.debug { "Saving group ${group.key}" }
-
-        return repository.save(group)
-    }
-
-    suspend fun findByIdOrNull(id: ObjectId): GroupDocument? {
-        logger.debug { "Fining group by ID: $id" }
-
-        return repository.findById(id)
     }
 
     suspend fun findByKey(key: String): GroupDocument {
@@ -91,10 +88,6 @@ class GroupService(
         return repository.existsByKey(key)
     }
 
-    suspend fun findById(id: ObjectId): GroupDocument {
-        return findByIdOrNull(id) ?: throw DocumentNotFoundException("No group with ID $id found")
-    }
-
     suspend fun update(key: String, req: UpdateGroupRequest): GroupDocument {
         logger.debug { "Updating group with key \"$key\"" }
 
@@ -103,7 +96,27 @@ class GroupService(
         val group = findByKeyOrNull(key)
             ?: throw DocumentNotFoundException("No group with key \"$key\" found")
 
+        req.translationsToDelete.forEach { locale -> group.translations.remove(locale) }
         group.translations.putAll(req.translations)
+
+        if (!group.translations.containsKey(appProperties.locale))
+            throw InvalidGroupTranslationException("Failed to update group: default locale ${appProperties.locale} is not contained in translations")
+
+        if (key == req.key || req.key == null) return save(group)
+
+        if (existsByKey(req.key)) throw GroupKeyExistsException("Failed to update group: a group with key ${req.key} already exists")
+
+        group.key = req.key
+
+        userService.findAllByGroupContaining(key)
+            .asFlux()
+            .map { user ->
+                user.groups.remove(key)
+                user.groups.add(req.key)
+                user
+            }
+            .buffer(1000)
+            .collect { users -> userService.saveAll(users) }
 
         return save(group)
     }
@@ -112,6 +125,17 @@ class GroupService(
         logger.debug { "Deleting group with key \"$key\"" }
 
         authorizationService.requireRole(Role.ADMIN)
+
+        if (!existsByKey(key)) throw DocumentNotFoundException("No group with key \"$key\" found")
+
+        userService.findAllByGroupContaining(key)
+            .asFlux()
+            .map { user ->
+                user.groups.remove(key)
+                user
+            }
+            .buffer(1000)
+            .collect { users -> userService.saveAll(users) }
 
         return repository.deleteByKey(key)
     }
